backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
neo4j==5.16.0
spacy==3.7.2
celery==5.3.6
redis==5.0.1
python-dotenv==1.0.1
httpx==0.26.0
pydantic==2.6.0
pydantic-settings==2.1.0
sentence-transformers==2.3.1
openai==1.12.0
